(GH-3) implemented multilang substantive substitution

+++ Codebase/equaly/src/main/java/org/hwr/equaly/model/Fragment.java
@@ -3,9 +3,9 @@ package org.hwr.equaly.model;
 public class Fragment {
     public String token;
     public String tag;
-    public long index;
+    public int index;
 
-    public Fragment(String token, String tag, long index) {
+    public Fragment(String token, String tag, int index) {
         this.token = token;
         this.tag = tag;
         this.index = index;
